+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # _Sharly Chess_ release notes
 
+## Tournaments
+
+- Exclusion of players (or teams) who have not played 50% of the games (or matches) in accordance with FIDE 6.6 in round robins (5.1.0)
+
 ## Pairings
 
 - Ask if a round should be marked as completed for individual round robins (5.1.0)